Factorize use_dict_output fixture to avoid duplicated code
