+++ deepinv/tests/conftest.py
@@ -7,6 +7,7 @@ from deepinv.utils import DownloadError
 from dummy import DummyCircles
 
 import importlib
+import contextlib
 
 # Tag stored on a TestReport's ``user_properties`` when we reclassify a
 # download failure as a skip. We attach it to the report (rather than to
@@ -189,6 +190,22 @@ def non_blocking_plots():
         # Restore the original backend
         matplotlib.use(original_backend, force=True)
         importlib.reload(plt)
+
+
+# The following fixture is used to catch deprecation warnings and test the deprecated tuple format for dataset outputs
+@pytest.fixture
+def use_dict_output(request):
+    # Catch deprecation warnings from previous dataset format
+    _use_dict_output = request.param
+    with (
+        pytest.warns(
+            DeprecationWarning,
+            match="The tuple format for dataset outputs is deprecated",
+        )
+        if not _use_dict_output
+        else contextlib.nullcontext()
+    ):
+        yield _use_dict_output
 
 
 # Certain tests are particularly slow and make for a large part of
